shortrfc3339: refactor into new package
This seems to be a common need, so moving it into its own space.
