+++ tariff/ngeso/api.go
@@ -6,6 +6,7 @@ import (
 	"errors"
 	"fmt"
 	"github.com/evcc-io/evcc/util/request"
+	"github.com/evcc-io/evcc/util/shortrfc3339"
 	"time"
 )
 
@@ -122,9 +123,9 @@ func (r NationalIntensityResult) Results() []CarbonIntensityForecastEntry {
 }
 
 type CarbonIntensityForecastEntry struct {
-	ValidityStart shortRFC3339Timestamp `json:"from"`
-	ValidityEnd   shortRFC3339Timestamp `json:"to"`
-	Intensity     CarbonIntensity       `json:"intensity"`
+	ValidityStart shortrfc3339.ShortRFC3339Timestamp `json:"from"`
+	ValidityEnd   shortrfc3339.ShortRFC3339Timestamp `json:"to"`
+	Intensity     CarbonIntensity                    `json:"intensity"`
 }
 
 type CarbonIntensity struct {
